pyproject.toml
[project]
name = "feishu-know-llm"
version = "0.1.0"
description = "AI-powered Feishu knowledge base manager using Claude"
requires-python = ">=3.10"
dependencies = [
    "lark-oapi>=1.4.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["ruff", "pytest"]
